move generate_code to utils
